Start gluing our system to Brig.
There are a couple of things we do to sit within Brig, the big one is where we add our command context to the String Reader. We do this to allow us to pass the context to parsers without mixins and further headaches with trying to make things painless for mods that use Brig directly.

This isn't complete, but it's a good way to what we're expecting. Registrars and parsers will come last.

Impl for CommandTreeBuilders

Currently missing the actual translation into Json - that's later

Add final modifiers, CommandCause is now a Subject and MessageReceiever.

Add interface to get CommandSources from the Cause easily.

Adding the CommandTreeBuilder scaffolding for custom frameworks

Once this is done, it's on to our own framework

More supporting code to enable Brig support.

Mostly need to do the grunt work - support our parameters now.

Starting to write class that builds the brig tree.

Starting complex parameters and how they will translate to the command tree

Add ClientCompletionType