Merge in SpongeVanilla
The key with this current build setup is that SpongeVanilla
now must be imported separately into an IDE by selecting the
vanilla/build.gradle.kts to import. It will resolve to building
SpongeCommon as a "child project", but otherwise, this allows
for SpongeVanilla to happily exist within the general repository
of SpongeCommon, without SpongeCommon building out the
SpongeVanilla in the case of the SpongeForge project setup.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>