Replace Sponge with PhaseTracker to access the CauseStackManager
It is an intrinsic part of the PhaseTracker that will be the entrypoint to handle
access for both PhaseTracking and CauseStack management. Since the PhaseTracker already
handles both SERVER and CLIENT thread setting, the access for other threads will be
equally checked by the PhaseTracker prior to providing a valid per-thread instance of
the CauseStackManager.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>