Allow to use a field cache for server and client PooledContexts
All other cases of 'other threads' pooled contexts will use the PhaseTracker's dedicated pool.
The uniqueness of the Server and Client threaded pools are enough that warrant a usage of caching
the latest singular state contexts, so we can support doing so. Being that the fields are read only
on validation that the thread is the same as their target, this should be threadsafe.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>