Enable World API based mixins
It is unclear whether the generic type erased methods are going to resolve
correctly or not. I have take a look at some of the emitted bytecode, but
could not verify that it is satisfactory enough to provide the mixins as
is without resorting to a bug submission to Mixin itself, or write some
default method handles that soft implement the type erased generic methods
like ChunkVolume.getChunk etc.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>