Replace UserDiscoverer with stripped back ServerUserProvider.
The idea of the UserManager is to manage User objects, not GameProfiles. While we do some operations with GameProfiles, they mirror what Users we think have been created.

The file system watching is in, but all GameProfile caching is delegated to the PlayerProfileCache, which we just get profiles from when required.

* Save the username cache json when we save the worlds
* Add stream methods to GameProfileCache
* Create UUID only GameProfiles for files that don't have an associated profile during the stream methods