Fix Server not being available when leaving SinglePlayer.
There is a timing issue that happens in the Minecraft issue where, along
the way, it checks to see if the ref on Minecraft for IntegratedServer
is null and if so it commences the shutdown process. I introduce a temp
workaround that keeps that ref for our usage throughout the shutdown
process but allows the engine to continue thinking it is gone.

This is the better compromise to me as it is minimal changes to the
Vanilla client, retains what the engine does, and keeps the API working.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>