Remove the concept of saving to config when setting property value.
This is more for zml but the insanity of stable-7 isn't going to
continue. When the config goes to save (which I believe is changing to
shutdown only or if a command is ran) then we would write values.
Otherwise, the values are to remain ON WorldInfo and the data is set on
the adapter before the adapter is told to save.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>