Implement new neighbor notification effect transaction recording
This has the side effect of eliminating the need for a NeighborNotificationState since we no longer
switch in to new PhaseContexts for neighbor events. NeighborNotification as a transaction will suit
enough of the needs to be able to explicitly record related transactions while reducing the overall
need to enter a new stack element to the JVM stack for processing.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>