Implement BlockTransaction cancellation and restoration
With the amount of events generated, we now can safely attempt
to iterate and cancel transactions. Relying on the ability to
link events to their transactions, we now also link the parent
or deciding transaction that if the event/event's transactions
are cancelled, the nested/children transactions will be marked
as cancelled, before their events are thrown. This allows for
snapshot restores to work again.

There's the caveat that without entity spawn tracking or
scheduled update event tracking, some interactions may still
proceed. When those are correctly tracked and correlated to
their respective transactions, we'll be able to see the full
transaction cancellation to take place.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>