Work around a client issue preventing some client completions from showing.
This is a complicated client side race condition and a complicated server side fix, thus I will try to give it some justice here.

When a player connects to the server, Minecraft creates a command tree to send to the client. We add our commands to this command tree, so that rich tab completion is available.

In order to enable custom elements, we have to specify that a parameter has "custom suggestions". This tells the client that they will need to ask us for suggestions. Minecraft discovers this by iterating over the command tree, parsing the command with the information it has, until it reaches a point where it can parse no futher. At this point, Minecraft looks at the nodes that could be next, and builds up suggestions based on that. In the case that a node requires custom suggestions, Minecraft will ask the server for the suggestions.

However, this breaks when more than one node is discovered that requires a server side completion. Minecraft can only handle one tab completion packet per player at a time. If two or more nodes are discovered requiring custom suggestions, that many tab complete requests are sent, each time creating a future that it will fulfill. However, each time a request is sent, any previous packet that was sent has its future discarded, so the server will calculate the suggestions n times when one would suffice.

While that's bad enough, the code that discovers the nodes returns a CompletableFuture. All the CompleteableFutures are gathered up in an "allOf" call and then chained to "thenRun" which processes the suggestions and then passes it onto the client for rendering. Unfortunately "allOf" returns a future which contains an exception if any of the futures are cancelled, meaning the completion will not be displayed.

The way around this is to manipulate the suggestion tree such that only one "custom suggestion" marker exists amongst siblings of a tree, as we'll return all node's suggestions anyway. That is what is done here.

* Also, smarten up suggestions a little by ignoring default nodes altogether.