Add PrepareBlockDropsTransaction for the quirks of Minecraft
Effectively, Minecraft requests that a block summons its drops
before the block is broken, so the effective way to capture these
drops and relate them to a block transaction after the fact
(in an event) is to make a "PreparedBlock" transaction and utilize
the parent-child-effect relationship. This works well with most all
cases that allow for the ChangeBlockEvent to be cancelled before the
SpawnEntityEvent is thrown for the block drops.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>