Add EntityPerformingDropsTransaction and related effect
This introduces a similar styled Transaction to be recorded when
an Entity is performing "death drop" like operations in the world
without expressly being ticked, cases are such as death update
ticks, entity deaths without an attack being performed, etc. The
intention of these recordings is to be able to associate not just
the entity spawns/items dropped, but also associate the block
changes that may occur by nature of the Transactor's parent-child-effect
handling.

This does add a shortcircuit on batching transactions for processing
since we do need to throw a HarvestEntityEvent individually for each
Entity being harvested.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>