Add VolumeStream implementation based on a Stream
The advantage of this implementation is that it is left to the
"supplier" of the elements to provide a stream of data. As it is
possible that a Stream could become out of date or out of sync, the
stream functionality is still retained throughout. In the instance
of implementing a BlockEntity stream in a World, there's the caveat
that we may need to decide on a few optional parameters to define
characteristics such as:
- Chunks loaded vs unloaded vs ungenerated
- Range restrictions
- Validation of tile entity positions after the position set is created
- Estimation of the stream size

It is my belief that performing a volume stream on a World would
intrinsically rely on some chunks being loaded and available, but this
may not be an expected behavior for plugin authors.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>