Introduce MemoizedSupplier for Supplier memoization
A Supplier normally would evaluate every single time being called, but in
some cases, we'd rather be able to memoize the value generated from the
Supplier after the first call. This is not meant to be leaked beyond memoizing
global objects that are referentially transparent themselves (like an object
being comprised of functions/suppliers that rely on other global variables).

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>