Reintroduce Blockstate(De)Serializer
Traditionally, Mojang is appearing to want to store a BlockState as a map to better
suit DFU's ability to update a state's values and properties. This map based
representation however, is not ideal to being used for storage with a palette that
retains the integer to string like mapping internally. It may well be possible to
take advantage of DFU in the future, but at this time, a less breakage of the
Schematic Specification is required.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>