Add item drops to EffectResult
This will be useful for calculating drops with relations throughout
block changes and entity changes occurring where we may be wanting
to throw a "pre" event. This has the direct advantage of
calculating the drops by loot table in various areas before performing
the spawns in a later set of transactions.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>