Add the ability to construct a WorldPipeline without processing
This will serve as a handy tool to adding extra effects after the fact of a block
change. Ideally, we would not be performing additional effects, but this is minecraft.
We have the unfortunate aspect that some effects (such as loot drops) are performed
separately/after a block change has been made.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>