Proxy PhaseState method access to avoid raw typing
Mostly a QOL change, as I've been thinking about it, the proxy access by
PhaseContext is handy, and can be kept tidy in PhaseStateProxy.

Demonstration of changes can be seen in ContainerUtil:
```diff
-  final IPhaseState<@NonNull ?> currentState = context.state;
-  if (((IPhaseState) currentState).doesBlockEventTracking(context)) {
+  if (context.doesBlockEventTracking()) {
```

I've not given thought to profile the difference as it is replacing an
explicit raw type cast to a method invocation by proxy. Seeing how it's
already relatively small, I can't imagine this would be an expensive
change.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>