Implement the Ticket Manager
Some notes: Tickets need to have a level of 33 or lower to generate a full chunk - that's what we're interested in here so that's our "minimum" level (meaning a maximum distance of 33 for each ticket). Further, in order for a chunk to be "fully loaded", a plugin should register a ticket with a reasonable lifetime - probably PORTAL in this case.