Fix OOM when creating paginations.
SpongeAdventure.asVanilla() creates an AdventureTextComponent, but this wasn't accounted for here - meaning "PaginationCalculator#getWidth()" always returned zero, and we then did a divide by zero combined with a GenericMath.floor call, meaning we got Integer.MAX_VALUE copies of the padding and no free memory.

Thus, we grab the vanilla component from the bridge, which is of the types that getWidth() expects them to be. This might not be the best way to fix it, but it works until the Adventure guys do their stuff (if anything).