Work around error when calculating permission nodes for commands.
This isn't really the complete fix for the issue, as "dangling nodes" are being created in some form. More investigation is needed - it might be that commands are being re-registered when datapacks are re-established without cleaning up first. However, this should stop exceptions occurring, instead producing a warning when it happens.