Strip non-minecraft dimensions from saving to world gen settings.
This prevents world data from loading when someone has removed a world template. Vanilla spec actually stores their worlds this way to defend against the level and sublevels being affected by changes to the server.properties (mainly if the seed changes).

In a future commit, I will write the Vanilla worlds to file as a template and have the file being the ultimate authority on what is read as properties.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>