Remove accessors from Vanilla platform.
This is a rule I am putting into place for our platform projects that unless the platform needs an accessor (for example, if we transform a Forge class we can actually do so to access a field not accessible, that cannot go into common of course), the platform will not run an accessors sourceset. This is a QOL change as it keeps all of our accessors largely together and cuts down on a sourceset to process during compilation.

This change is done for SpongeVanilla, SpongeForge is left alone as it is not being worked on currently.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>