Properly shutdown the installer's child process (#3260)
This is the proper solution for #3251. Without this change, shutting down the server only works by pressing ctrl-c while having a terminal open with the stdin of the child process (actual MC server), since the child is started with `inheritIO()`. When the stdin is not kept open, which would be normal for a swarm deployment, before this commit there was actually no way of cleanly shutting down the server, because PID1 received the INT/TERM signal, started its shutdown hook which then waits forever for the child to exit, since the child does not know it's supposed to exit. Now it does, I tested this in a few scenarios.