Bump guice to 5.0.1
Because of our custom SpongeInjectionPoint and use of Guava 21, we have
to use a fork of Guice 5.0.1 that restores API removed in Guice 5
without a suitable alternative.

Upstream issue: https://github.com/google/guice/issues/1121

This newer version of Guice is needed to maintain compatibility with
newer Java versions which plugin developers may want to use.