Add children instead of redirect where possible due to a bug in Brigadier
Rather, this looks like an intentional decision which actually causes huge problems with the client. A lot of our changes are designed to work around Brig but we can't affect the vanilla client. The problem is that the equality contract for nodes don't account for redirects, so if you have two literal nodes with the same name but two different redirects (or none at all), Brig (incorrectly) assumes they are the same and the client just merges them, so the client completion for one command may appear on another.

Redirects would be better simply to reduce the tree size, so if/when fixed on the client we should revert this.

This may solve #3349 and is also the same issue with command flags, but because of the way flags work we need the redirects and therefore needs a better workaround (or for Mojang to fix this)

Also fixes compile error introdueced by https://github.com/SpongePowered/Sponge/commit/5e202d4fb24db5eb194658d9f9c85920770412d3 as best as I can.