Work around Vanilla bug with components in commands
Basically an off by one error meaning that after parsing a Component Json object, a command's StringReader has been advanced to the next non-whitespace character. This mixin detects this situation and undoes it to allow for parameters past such json. While the client may still display that the parameter can't be parsed unless two spaces are supplied, this parameter will now work either way - if a user uses a double space so completions work, that is fine, if a user doesn't, that is also fine.

We're just going to have to live with the bug on the client, we can't do anything about it on the vanilla side (but this has been fixed if you run the client with Sponge)

Fixes #3361