Fix issues with nested transactions avoiding events
In some cases of how block drops are performed,
the child transactions were being ignored. This
does a few behavioral changes based on the code
introduced by optionality on worlds to now having
optionality of the event itself contributing to
child transaction's events.

Furthermore, an added change to iterate in a
reverse order of transactions logged to match a
removal of a BlockEntity is done to better match
more recent transactions that would have otherwise
represented the removal of that block entity.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>