Add functionality to absorb drop preparations to tail
For block changes that occur in survival, it's
often done in a way that the block drops are
performed separately from the block being broken,
leading to a separation of the "PrepareDrops" effect
from the ChangeBlock transaction. This makes it
relatively easy to abosrb a prepare drops into the
tail for those cases.

It is not tested or guaranteed this will function
for "multi-block" mining in survival, however that
is done on a mod basis. We could investigate at a
later time whether a crawling iteration to
associate drops on previously recorded block
changes.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>