Fix ignoreMovement causing leaked block changes
@Zidane will probably laugh at me for it, but, this is
the simplest way to figure out whether we still need to
run the tracking logic around a packet if the player is
around a block that has overridden entity inside logic.

In this way, we are able to avoid leaking changes from
players standing idle in place potentially triggering
things like pressure plates.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>