Don't load neighbour chunks during load/unload
This might not be the complete fix to an issue that was exposed by 429de3b0261bcd066fded34bb7a97677fc545171, but neighbour chunks should not be loaded to set fast access flags - they should only be there if both chunks are loaded in the first place (if the neighbour chunk is subsequently loaded then the links will be made)

The issue that the linked commit seemed to cause was a deadlock caused by a chunk load being triggered within the chunk load. There is probably a bigger issue that needs fixing here, but this enables those who want to test the build to get back underway.