Re-think how #sponge data goes into the codecs
By default, Vanilla (when doing a merge of codecs) will place the
resultant merger under a "value" key by way of logic in
KeyDispatchCodec. Which violates Vanilla has their map data was being
moved around.

As a workaround, the new "CombineMapsCodec" exists in the Vanilla
system's bounds and appropriately handles the merge on our own.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>