Refactor app launch to remove plugin-spi requirement
Mostly remove it, not completely. Should allow Forge and other platforms
to easier handle it. While the particular platform will need to
implement a PluginLanguageService, it can just be a wrapper around that
platform's mod system (and not trying to make plugin-spi to hook into
that system during the bootstrap)

Signed-off-by: Chris Sanders <zidane@spongepowered.org>