vanilla: Add dependency resolution
This ensures plugins are loaded in an appropriate order, so plugins without dependencies are loaded first, and the algorithm recursively determines the load order based on satisifed dependencies. BEFORE dependencies can be supported by uncommenting the code within the TODOs in DependencyResolver, it treats thouse as "optional after" deps on the target -- though will still fail if the dependency is not there and is not optional.

If we have optional dependencies, there is a chance that a cyclic chain may occur. For optional deps, however, they should be able to handle a dependent plugin being loaded before them because they are already "optional". Thus, if we find that no plugins were loaded in a pass and but there were optional dependencies that could be ignored in that scenario, we attempt to load plugins that only have their compulsory dependencies loaded.