Update and resolve usage of IPhaseState.isCollision
Historically this function was catered to ticking
states and when we specifically entered entity
collision detection during an entity's tick. In
various refactors, the meaning got lost over time,
and now we can assuredly resolve that the function
is more correctly named allowsEntityCollisionEvents
to be thrown, which otherwise should not be thrown
in an6y other state.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>