Slightly refactor the login process to try to catch exceptions
* Reduced object creation.
* Handle client specific checks that were missed.

Note that there seems to be an issue on the client where cancelling the login will cause the client to hang. This does not happen on the server. Will look at trying to move the login event to where the disconnection normally occurs.

See #3469 - this may resolve the issue if the problem was actually due to an exception