Temporary fix finding Users by UUID
If a Player has played before but not in this server session, Sponge code assumes that Vanilla has the game profile cached but that does not happen until the Player joins once. As seen with the TODO, this is by far not the correct fix. We need to re-think the UserManager.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>