Refactor how Users work.
* Split UserData from its view so that we don't have to worry about user data when a player is online.
* Don't load users in the PhaseTracker
* Ensure we're saving the game profile cache, and make sure we don't wipe it on startup

We also prevent players with broken files from logging in to protect their data.

Vanilla just brushes it all under the carpet and effectively deletes the data. We make sure we keep the data and stop the login process, so that server owners can be alerted to data failures.

We also copy the corrupted data for additional protection.