forge: Workaround class change errors with super defaults
In the forge workspace, if a Mixin implements an interface with default
methods and in those overrides, if they call to the default super
method, an incompatible class change with "methodref must be constant"
will occur. This commit works around this by moving the logic out to a
static processing class and invoking those methods manually in
ItemStackMixin as was happening previously.

A nasty workaround that hopefully can be reverted one day.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>