Always broadcast changes on the end of any Packet
This avoids some persistence cases where a variety of slot changes ends
up being sent to the client in the next tick, losing any sort of cause
relation for the desired events. This has the side effect that we're now
always transacting inventory events if and only if the inventory has any
changes actually necessary.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>