Always log accidental spawn records
Many of these inventory states do not expressly allow for entities to be
captured, but the corner case that an erroneous call from a transaction
in these particular states will result in a warning log. The
ramifications of the records are nil for gameplay, but mean that the
entities spawned may not have an event thrown, and therefor plugins may
not be aware of them.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>