Properly record Menu opening
Menus are being created and then set on the player after the player is
added as a container listener. This unfortunately means that we have to
record the transaction for the player, then after the menu is being set,
record the new container on the specific transaction.
This is a tad brittle in the sense that we're not always sure that an
OpenInventory is being performed (since the container menu can be null),
but the resulting effect makes for an empty event optional, which turns
into a no-operation during processing.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>