Move Transactional business logic to interfaces
The goal with this commit is to better organize the different behaviors
of GameTransactions that now exist and likely will increase as more is
refactored/cleaned up. The ultimate goal of this piece is to better
improve the inventory transaction management and have a less brittle
nature to their recording and sanity.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>