Introduce proper GameTransaction iterators
Unit tested iteration ordering and processing, both in isolation and
with the TransactionalCaptureSupplier in play. The benefit of these
iterators is the guarantee now by contract as to "how" the transactions
are iterated through when attempting to merge/absorb a transaction into
any existing effects/transactions.

There's a huge speed boost for noisy transactions with
TransactionFlow.canBeAbsorbed, which enables some fail-fast and iterator
avoidance for most cases.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>