Encapsulate and defer initialization of snapshots
Transactions can now have some lazy initialization by virtue of the
StatefulTransaction. Slowly the transaction can be lazily initialized
with at-least-requirements to be potentially absorbed by a parent
transaction, but not completely empty of data that the deferral renders
data to be unusable. The end goal of a recorded transaction is that it
knows its final state so-long-as it has populated any of it's non-final
fields.

Some inventory transactions intrinsically do not have the luxury of
deferred initialization and instead must await for future transactions
to be accepted down the line.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>