Wrap around event listener registration
Avoids game crashes for seemingly client sided methods.
Client sided methods can be included in class bytecode, but
ultimately it's up to the developer to exclude it properly. In our case,
we don't have control over it normally, so we'll have to expect any
client side code to trigger some RuntimeExceptions.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>