ASMify ReflectionUtil
The end goal is to replace usage of Java Reflection API that can be prone
to classloading exceptions/errors that end up crashing or throwing errors.
The major problem with reflection is the lack of clear control over
ClassLoaders and their ability to throw any arbitrary error. In retrospect,
we can now safely scan entire classes for specific methods that we care
about without triggering these classes to be loaded. Likewise, as a test,
we can assure that a defined class with an impossible to use method will
still be correctly covered for our purposes.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>