ASMify SpongeEventManager
This was brutal awakening of how much reflection is relied on to create all sorts
of classes. Fortunately, the brunt of the changes are to do with figuring out how
to use ASM visitors to do the discovery work. It is now unit tested to ensure
classes that can be triggered with reflection for failure, will not fail.
An advantage of this is now we are assured the only class loading failures would
be breaking changes to our own API classes.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>