During loading the level, gate querying for the stem's dimension type
This is considered a criical crash condition however the previous error
gave no real information as to why (it only stated a type was not found
and printed just the hash of the object). While there is little
information available for user-facing debug at this point, we can at
least provide the particular level that is being loaded which should
hopefully serve as a clue.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>