vanilla: Force SLF4J to rescan for Log4J service providers.
For some reason, SLF4J was not using the Log4J provider all the time. Adding the 1.8 adapter to the installer caused it to work sometimes, as did putting in a breakpoint and trying to debug it... Instead, to get this working (at least), use reflection to cause slf4j to rescan for the Log4J provider and re-enable logging during server start - which also causes all subsitute loggers to flush to Log4J, meaning no logging is lost.

There probably needs to be a better fix but this allows server owners to see their logs again in the short term.

Also removed the explicit Jansi dep, as that's included in the Jline dependency.

Fixes #3696