Fixes to serialization behavior
We were marking world unload as manual save, it should be explicitly asked for.

The documentation of ServerWorld#save states that it should return false on SerializationBehavior.NONE.
However still call the events to be consistent with how it works for other sources.

Make sure we set the isManualSave to false in case the save event is cancelled. Otherwise the next call would be errously marked as manual save.

Fixed that we did not do automatic saves on saves marked as manual, we should always on automatic behaviour.