Prevent invalid status responses with empty/dummy player UUIDs (#3840)
Right now SpongeGameProfile.toMcProfile() converts all SpongeAPI GameProfiles
with empty UUIDs (= new UUID(0, 0)) to authlib GameProfiles with a null UUID.

This is fine in most cases but problematic for the sample/player list in
the status response. Empty UUIDs can be used there to create dummy entries,
to display custom messages in the server list instead of the online players.
(See e.g. https://github.com/Minecrell/ServerListPlus/wiki/Status-Configuration#player-hover-messages)

The Minecraft client completely rejects status responses that contain
GameProfiles with null UUIDs and/or null names. This is why it is
important that the empty UUID is not converted to a null UUID.

Prevent sending invalid responses to the client by making sure both
UUID and name are non-null in the player samples of the status response.

This fixes a regression in API 10 when using the ServerListPlus plugin
on Sponge. The problem was likely introduced in some GameProfile refactoring
(API 7 worked fine).