convert arrays from NBT to JSON.
Arrays are not an Iterable instance like lists are.

Also accept long arrays when translating NBT.

This fixes plugins having errors when trying to convert NBT with int/long/byte arrays to JSON.
Which usually happens with firework items (the Colors and FadeColors are int arrays).

However upon converting this JSON back to NBT, it will create lists of ints instead of int arrays etc.