Fix minor issues with show_entity hover in messages (#3908)
* actually use show entity hover event type, instead of using name in the wrong place

* don't attempt to getString of null resources when showing entity type in hover

EntityList does not contain all possible entity types. It does not contain entries for:
FishingHook, lightning, Weather, Player, or ComplexPart.
Trying to use a hover show entity of one of these types would cause a null pointer exception.
And before the previous commit the type was actually the name.
Which caused showing entities with a name of "Player" or one of the other types mentioned above to cause this error.
Preventing the message from being sent.