Hacky replace of server-level registries.
On the SpongeVanilla client, it is full intentional for client saves to be entered and exited. Upon enter, the Vanilla client bootstraps the server and we load in server-level registries and register them. On a dedicated server, this process occurs once. However, it repeats on the client. When it does, the code rightfully crashes. This commit is a very dirty temp fix for this problem by allowing certain registries to bypass the freeze and replace the instance. The Vanilla registry system handled this challenge via the RegistryLayer system. A long term fix would be to change the Sponge registry implementation to sit on top of Mojang's work.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>