Adjust PhaseState pooling
This is a correctness fix as we should not store the states
before they are properly closed and the state has been cleared.

We also didn't clear or return states to the pool that
didn't provide modifiers and could leak resources for
long periods of time until they were used again.

As added bonus we now pool more states and removes some branching.